[IMP] tools: use `DotDict` only in tests
The `DotDict` class can produce unexpected behaviour
if used incorrectly.
This commit moves this class so that it is only used in tests.

task-4822364
